refactor(032-ship-stage): split workflow.ts into query + advance + shared (T001)
Extract the mutating advance/link/redesign verbs into
src/subcommands/workflow-advance.ts and the shared resolution helpers
(resolve, phaseById, forwardTransition, failUsage) into workflow-shared.ts,
keeping the read-only query verbs + dispatcher in workflow.ts. Pure
behavior-preserving refactor (R7) so each file stays <=500 lines before the
ship/backstop wiring lands. T002: makeWorkflowFixture already covers an
on-disk install with a govern convergence record + configurable node status.

All 163 workflow tests green; no behavior change.
